app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.0.1
certifi==2024.8.30
click==8.1.7
coverage==7.6.1
fastapi==0.112.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
mysql-connector-python==9.0.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pytest==8.3.2
pytest-cov==5.0.0
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
